+++ src/database/postgres/model_user.rs
@@ -127,7 +127,6 @@ where
             .await
             .map_err(|_| ApiError::Internal(S!("jar")))?;
         let state = ApiState::from_ref(state);
-
         if let Some(ulid) = get_cookie_ulid(&state, &jar)
             && let Some(user) = RedisSession::get(&state.redis, &state.postgres, &ulid).await?
         {