+++ src/photo_convertor.rs
@@ -19,9 +19,9 @@ pub struct PhotoLocationEnv {
 impl PhotoLocationEnv {
     pub fn new(app_env: &AppEnv) -> Self {
         Self {
-            converted: C!(app_env.location_photo_converted),
-            original: C!(app_env.location_photo_original),
-            watermark: C!(app_env.location_watermark),
+            converted: C!(app_env.location.photo_converted),
+            original: C!(app_env.location.photo_original),
+            watermark: C!(app_env.location.watermark),
         }
     }
 
@@ -107,7 +107,7 @@ impl PhotoConvertor {
             let img =
                 image::load_from_memory_with_format(&original_bytes, image::ImageFormat::Jpeg)?;
 
-            let mut converted_img = img.resize(1000, 1000, image::imageops::FilterType::Nearest);
+            let mut converted_img = img.resize(1000, 1000, image::imageops::FilterType::Lanczos3);
             let watermark = image::open(location_watermark)?;
             let watermark_x = i64::from(converted_img.width() - watermark.width() - 4);
             let watermark_y = i64::from(converted_img.height() - watermark.height() - 4);
@@ -120,7 +120,7 @@ impl PhotoConvertor {
                     ))),
                 ))
             })?;
-            let webp = encoder.encode(75.0);
+            let webp = encoder.encode(95.0);
             Ok(webp.to_vec())
         })
         .await??;