+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+# <a href='https://github.com/mrjackwills/mealpedant_api/releases/tag/v2.2.0'>v2.2.0</a>
+### 2026-08-27
+
+### Chores
++ .gitignore updated, [7a354cb9](https://github.com/mrjackwills/mealpedant_api/commit/7a354cb9f8e95260e609cf160865f6ffee8a3ba2)
++ dependencies updated, [a20c945f](https://github.com/mrjackwills/mealpedant_api/commit/a20c945f69fd2e1a322c9694d1915c2485d56aba)
++ Rust version bump, [82a93051](https://github.com/mrjackwills/mealpedant_api/commit/82a93051a378d4235ec1d823ce36356c96f50409)
++ GitHub workflow updated, [4f6995ab](https://github.com/mrjackwills/mealpedant_api/commit/4f6995ab275027895d6353cf1ab5506aa5bb4332)
+
+### Features
++ log rotation, domain_env, [3fd6676d](https://github.com/mrjackwills/mealpedant_api/commit/3fd6676d8446216b337f5e6285a77c156d583328)
++ require auth for 2FA backup codes, [c6a91204](https://github.com/mrjackwills/mealpedant_api/commit/c6a91204eb134066e9025a6e87858fdff5c42495)
++ require password on 2fa alwaysrequired, [cdc6dc1e](https://github.com/mrjackwills/mealpedant_api/commit/cdc6dc1e62a66e6df4494b378510631345dfb5eb)
++ update postgres container settings, [752d1fb9](https://github.com/mrjackwills/mealpedant_api/commit/752d1fb962d2bfc2f9c8fa39c737d1d4b44e5024)
++ update redis container settings, [56c6bf4b](https://github.com/mrjackwills/mealpedant_api/commit/56c6bf4b82f9aab2b11ff525fe610b829db81d29), [5c071261](https://github.com/mrjackwills/mealpedant_api/commit/5c071261f9feb09381dd43a83a2c59cfd4a7299e)
++ use single DB connection pool(s), [f575fa70](https://github.com/mrjackwills/mealpedant_api/commit/f575fa70441fe3be592938b908fadf2127775f16)
++ improve webp quality, [94ed5590](https://github.com/mrjackwills/mealpedant_api/commit/94ed55904d9c850cbaf701a92df80973e3c886cb)
+
+### Fixes
++ add ttl to user session set, [2b649f08](https://github.com/mrjackwills/mealpedant_api/commit/2b649f0877dbd2ad61861fa45fc4bb577317b15d)
++ improve converted image, [fd1689af](https://github.com/mrjackwills/mealpedant_api/commit/fd1689afafb5c7787970d385d861dc5ed49618ab)
++ macro typo, [bb74552a](https://github.com/mrjackwills/mealpedant_api/commit/bb74552afa2e9d9094f60b1ff0b6f8e9bd047d08)
++ ratelimit, [c9688ae9](https://github.com/mrjackwills/mealpedant_api/commit/c9688ae945fcdfc416bdc408e16662280bbb6323)
+
+### Refactors
++ remove reduduant code, [4b447c45](https://github.com/mrjackwills/mealpedant_api/commit/4b447c4557e1506c015385687dce7803c5b35a51), [efc08faf](https://github.com/mrjackwills/mealpedant_api/commit/efc08fafb8846d6a5fa3cc1bfeb938525ca31c08)
++ auth code refactored, [33787f9f](https://github.com/mrjackwills/mealpedant_api/commit/33787f9f026f5dc3027394330a60636dbaef5843)
++ dead code removed, [1c0d95b3](https://github.com/mrjackwills/mealpedant_api/commit/1c0d95b32f0118fbfd1c9f34541e7e305673c80d)
+
 # <a href='https://github.com/mrjackwills/mealpedant_api/releases/tag/v2.1.4'>v2.1.4</a>
 ### 2026-07-16